refactor(cyclonedx): Move remaining Bom extensions to the reporter
In a recent change, the model mapping code has been moved out of the
reporter. The main concern of the remaining reporter code is writing
the report files. So, the reporter is a good place for the
corresponding functions.
